requirements.txt
google-generativeai
pvporcupine
openai-whisper
pyttsx3
python-dotenv
pyaudio
pyautogui